// src/app/shared/models/product-category.ts
export class ProductCategory {
    ID: number;
    Name: string;
    Alias: string;
    Description: string;
    ParentID: number;
    DisplayOrder: number;
    Image: string;
    HomeFlag: boolean;
    CreatedDate: string;
    CreatedBy: string;
    UpdatedDate: string;
    UpdatedBy: string;
    MetaKeyword: string;
    MetaDescription: string;
    Status: boolean;
}